API Routes secured

+++ server/controllers/loanController.js
@@ -3,17 +3,29 @@ const Loan = require("../models/loanModel");
 const mongoose = require("mongoose");
 const bot = require("../bot"); //Bot commands can be called directly in this script, and will work.
 
-// get all loans
+const getAuthStatus = (req, res, id) => {
+  if (!req.user) {
+    return res.status(401).send("Unauthorized");
+  }
+  if (req.user.id !== id) {
+    return res.status(403).send("Forbidden");
+  }
+  return null; // Return null if checks pass
+};
+
+/* get all loans
 const getLoans = async (req, res) => {
   const loans = await Loan.find({}).sort({ createdAt: -1 });
 
   res.status(200).json(loans);
 };
+*/
 
-// get single loan
+/* get single loan
 const getLoan = async (req, res) => {
   const { id } = req.params;
 
+  
   if (!mongoose.Types.ObjectId.isValid(id)) {
     return res.status(404).json({ error: "No such loan" });
   }
@@ -26,9 +38,17 @@ const getLoan = async (req, res) => {
 
   res.status(200).json(loan);
 };
+*/
+
 const getUsersLoans = async (req, res) => {
   try {
     const { id } = req.params;
+
+    const authStatus = getAuthStatus(req, res, id);
+
+    if (authStatus) {
+      return authStatus;
+    }
 
     const loans = await Loan.find({ "lender.lenderId": id }).sort({
       createdAt: -1,
@@ -44,6 +64,12 @@ const createLoan = async (req, res) => {
   // add doc to db
   try {
     const { loanName, guild, lender, borrowers } = req.body;
+
+    const authStatus = getAuthStatus(req, res, lender.lenderId);
+
+    if (authStatus) {
+      return authStatus;
+    }
 
     const loan = await Loan.create({
       loanName: loanName,
@@ -62,6 +88,12 @@ const updateLoan = async (req, res) => {
   try {
     const { loanName, guild, lender, borrowers } = req.body;
     const { id } = req.params; // Get loan ID from request params
+
+    const authStatus = getAuthStatus(req, res, lender.lenderId);
+
+    if (authStatus) {
+      return authStatus;
+    }
 
     // Find and update the loan
     const loan = await Loan.findByIdAndUpdate(
@@ -89,11 +121,18 @@ const deleteLoan = async (req, res) => {
     return res.status(404).json({ error: "No such loan" });
   }
 
-  const loan = await Loan.findOneAndDelete({ _id: id });
+  const loan = await Loan.findOne({ _id: id });
 
   if (!loan) {
     return res.status(400).json({ error: "No such loan" });
   }
+
+  const authStatus = getAuthStatus(req, res, loan.lender.lenderId);
+
+  if (authStatus) {
+    return authStatus;
+  }
+  await Loan.findOneAndDelete({ _id: id });
 
   res.status(200).json(loan);
 };
@@ -127,6 +166,13 @@ const removeBorrower = async (req, res) => {
 const remindLoan = async (req, res) => {
   try {
     const { id } = req.params;
+    const loan = await Loan.findOne({ _id: id });
+
+    const authStatus = getAuthStatus(req, res, loan.lender.lenderId);
+    if (authStatus) {
+      return authStatus;
+    }
+
     bot.remind(id);
     res.status(200);
   } catch (error) {
@@ -136,8 +182,8 @@ const remindLoan = async (req, res) => {
 };
 
 module.exports = {
-  getLoans,
-  getLoan,
+  //getLoans,
+  //getLoan,
   createLoan,
   deleteLoan,
   updateLoan,
